pyproject.toml
[project]
name = "resumeio-to-pdf"
version = "1.0.0"
description = "Download your resume from resume.io as PDF"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.0",
    "jinja2>=3.1.6",
    "pdf2docx>=0.5.6",
    "pillow>=11.3.0",
    "pypdf>=6.0.0",
    "pytesseract>=0.3.13",
    "uvicorn>=0.35.0",
]
[dependency-groups]
dev = [
    "black>=25.1.0",
    "ruff>=0.12.12",
]

[tool.black]
line-length = 120
verbose = true
preview = true

[tool.ruff]
line-length = 120
lint.select = [
    "F",  # pyflakes
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "I",  # isort
    "N",  # pep8-naming
    "COM",  # flake8-commas
    "Q",  # flake8-quotes
]
